core(gating): stabilize shared-state tests with serial_test\n\n- Fix flakiness by serializing tests in arw-core gating module\n- Add dev-dep serial_test\n\nevents(bus): clippy cleanup\n- Replace match loop with while-let in subscribe_filtered\n\notel: remove unused prelude import when otlp feature\n\ncore(wasm): remove duplicated cfg attribute per clippy\n\nconnector: avoid unreachable tail by running worker in task and waiting for Ctrl+C\n\nsvc(io): clarify OpenOptions (explicit truncate(false)) for lockfile\nsvc(events): remove unnecessary cast in SSE gap handler\n\ndocs: update tasks + status + roadmap (tests+lint)\n- scripts/tasks.sh now marks new tasks done with notes\n- scripts/docgen.sh regenerated status, tasks, and specs

+++ apps/arw-svc/src/main.rs
@@ -575,7 +575,7 @@ async fn events(
             ))
         }
         Err(BroadcastStreamRecvError::Lagged(n)) => {
-            bus_for_lag.note_lag(n as u64);
+            bus_for_lag.note_lag(n);
             let body = format!("{{\"skipped\":{}}}", n);
             tokio_stream::once(Ok::<Event, Infallible>(
                 Event::default().id("gap").event("Bus.Gap").data(body),
